Refactor the wiseagents.graphdb package
Fixes https://github.com/fjuma/wise-agents/issues/36
