+++ src/wiseagents/graphdb/lang_chain_wise_agent_graph_db.py
@@ -1,16 +1,36 @@
 from typing import Optional, List
 
 from langchain_community.graphs import Neo4jGraph
-from wiseagents.graphdb.Source import Source
-from wiseagents.graphdb.Entity import Entity
-from wiseagents.graphdb.GraphDocument import GraphDocument
-from wiseagents.graphdb.LangChainWiseAgentGraphDB import LangChainWiseAgentGraphDB
+from langchain_community.graphs.graph_document import (GraphDocument as LangChainGraphDocument,
+                                                       Relationship as LangChainRelationship, Node)
+from langchain_core.documents import Document as LangChainDocument
 
-from wiseagents.graphdb.Relationship import Relationship
+from .wise_agent_graph_db import Entity, Source, GraphDocument, Relationship, WiseAgentGraphDB
+
+
+class LangChainWiseAgentGraphDB(WiseAgentGraphDB):
+
+    def convert_to_lang_chain_node(self, entity: Entity) -> Node:
+        return Node(id=entity.id, type=entity.label, properties=entity.metadata)
+
+    def convert_to_lang_chain_relationship(self, relationship: Relationship) -> LangChainRelationship:
+        return LangChainRelationship(source=self.convert_to_lang_chain_node(relationship.source),
+                                     target=self.convert_to_lang_chain_node(relationship.target),
+                                     type=relationship.label,
+                                     properties=relationship.metadata)
+
+    def convert_to_lang_chain_graph_document(self, graph_document: GraphDocument) -> LangChainGraphDocument:
+        return LangChainGraphDocument(nodes=[self.convert_to_lang_chain_node(entity)
+                                             for entity in graph_document.entities],
+                                      relationships=[self.convert_to_lang_chain_relationship(relationship)
+                                                     for relationship in graph_document.relationships],
+                                      source=self.convert_to_lang_chain_document(graph_document.source))
+
+    def convert_to_lang_chain_document(self, source: Source) -> LangChainDocument:
+        return LangChainDocument(id=source.id, page_content=source.content, metadata=source.metadata)
 
 
 class Neo4jLangChainWiseAgentGraphDB(LangChainWiseAgentGraphDB):
-
     yaml_tag = u'!Neo4jLangChainWiseAgentGraphDB'
 
     def __init__(self, url: Optional[str] = None, refresh_graph_schema: Optional[bool] = True):
@@ -77,9 +97,3 @@ class Neo4jLangChainWiseAgentGraphDB(LangChainWiseAgentGraphDB):
         """
         if self.neo4j_graph_db is not None:
             self.neo4j_graph_db._driver.close()
-
-
-
-
-
-
